+++ Editor/Utils/TextureUtils.cs
@@ -105,13 +105,13 @@ namespace Unity.ProjectAuditor.Editor.Utils
         /// </summary>
         /// <param name="spriteAtlas">The Sprite Atlas to check.</param>
         /// <returns>The percent of empty space.</returns>
-        internal static int GetEmptySpacePerecentage(SpriteAtlas spriteAtlas)
+        internal static int GetEmptySpacePercentage(SpriteAtlas spriteAtlas)
         {
             var method = typeof(SpriteAtlasExtensions).GetMethod("GetPreviewTextures", BindingFlags.Static | BindingFlags.NonPublic);
             object obj = method.Invoke(null, new object[] { spriteAtlas });
             Texture2D[] textures = obj as Texture2D[];
 
-            if (textures == null)
+            if (textures == null || textures.Length == 0)
             {
                 Debug.LogError($"Could not load texture from {spriteAtlas.name}");
                 return 0;