+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+* Added Empty Sprite Atlas analyzer
+* Added Empty Atlas Texture analyzer
 
 ### Added
 
@@ -12,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Bumped analytics events version number from v1 to v2
 * Made many APIs internal rather than public.
 * Added properly-formatted comments to enable API documentation for all remaining public types and methods
+* Asset diagnostics IDs for Resources/StreamingAsset folder issues
 
 ## [0.9.4-preview] - 2023-03-27